refactor: further modularize code
